refactor(866): move Email/Notifications/Issues Hangfire jobs into their sections (G5 lane 5b-1)
Moves the four recurring-job types owned by Email, Notifications and Issues out of
src/Humans.Infrastructure/Jobs into their owning section projects, under Contracts/.
HangfireImmediateOutboxProcessor follows ProcessEmailOutboxJob for the same reason.

The premise that pinned them to Base was re-measured and is false: every job is
registered with RecurringJob.AddOrUpdate<T>(id, ...), the id is the stable key, and
AddOrUpdate rewrites the stored type string at every startup.

No job id changed, no registration site moved, and the retired array was not touched.

Refs nobodies-collective/Humans#866
