+++ src/Humans.Web/Extensions/Infrastructure/EmailInfrastructureExtensions.cs
@@ -1,7 +1,5 @@
 using Humans.Email.Contracts;
 using Humans.Infrastructure.Configuration;
-using Humans.Infrastructure.Jobs;
-using Humans.Infrastructure.Services;
 
 namespace Humans.Web.Extensions.Infrastructure;
 
@@ -10,9 +8,16 @@ namespace Humans.Web.Extensions.Infrastructure;
 /// binding — read by Auth's magic-link URL builder, Profiles' unsubscribe token provider,
 /// <c>SendReConsentReminderJob</c> and <c>SmtpHealthCheck</c>, so it is Base configuration
 /// the section is merely named after — plus the startup guard that Production must have
-/// SMTP configured, the Hangfire-backed <see cref="IImmediateOutboxProcessor"/> and the two
-/// recurring jobs (design §15 step 6b). Everything else moved into <c>Humans.Email</c>'s
+/// SMTP configured. Everything else moved into <c>Humans.Email</c>'s
 /// <c>Section.Register</c>.
+///
+/// <para>
+/// The three Hangfire registrations below still live here, but the <em>types</em> they name
+/// moved into <c>Humans.Email/Contracts/</c> at G5 lane 5b-1
+/// (nobodies-collective/Humans#866). Only the registration stayed: Shell references every
+/// section, so naming a section's concrete type here is free, and a section has no
+/// <c>ISection</c> discovery seam for recurring jobs yet.
+/// </para>
 /// </summary>
 internal static class EmailInfrastructureExtensions
 {
@@ -44,12 +49,12 @@ internal static class EmailInfrastructureExtensions
                 "Email SMTP configuration is required in production. Set Email:Host.");
         }
 
-        // Stays in Shell rather than the section: the implementation is Hangfire-backed and
-        // lives in Humans.Infrastructure, and Program.cs already treats the
-        // HangfireImmediateOutboxProcessor → IImmediateOutboxProcessor → OutboxEmailService
-        // chain as Shell's (see the Testing-environment Hangfire skip). The section registers
-        // OutboxEmailService as IEmailService and takes this as a constructor dependency, so
-        // ValidateOnBuild fails at startup without it.
+        // The registration stays in Shell even though the implementation moved into the
+        // section at G5 lane 5b-1: Program.cs treats the HangfireImmediateOutboxProcessor →
+        // IImmediateOutboxProcessor → OutboxEmailService chain as Shell's (see the
+        // Testing-environment Hangfire skip), and the section registers OutboxEmailService as
+        // IEmailService taking this as a constructor dependency, so ValidateOnBuild fails at
+        // startup without it.
         services.AddScoped<IImmediateOutboxProcessor, HangfireImmediateOutboxProcessor>();
 
         services.AddScoped<ProcessEmailOutboxJob>();
