+++ src/Humans.Web/Extensions/RecurringJobExtensions.cs
@@ -1,7 +1,10 @@
 using Hangfire;
 using Humans.Application.Configuration;
 using Humans.Application.Interfaces.GoogleIntegration;
+using Humans.Email.Contracts;
 using Humans.Infrastructure.Jobs;
+using Humans.Issues.Contracts;
+using Humans.Notifications.Contracts;
 
 namespace Humans.Web.Extensions;
