+++ src/Sections/Humans.Email/Contracts/CleanupEmailOutboxJob.cs
@@ -1,15 +1,19 @@
 using Hangfire;
 using Humans.Application.Interfaces;
-using Humans.Email.Contracts;
 using Microsoft.Extensions.Logging;
 
-namespace Humans.Infrastructure.Jobs;
+namespace Humans.Email.Contracts;
 
 /// <summary>
 /// Purges old sent messages from the email outbox. Runs weekly. The retention cutoff —
 /// <c>Email:OutboxRetentionDays</c> back from now — lives inside the Email section behind
 /// <see cref="IEmailOutboxRetention"/>; this job is the scheduler shim around it.
 /// </summary>
+/// <remarks>
+/// Moved out of <c>Humans.Infrastructure/Jobs</c> at G5 lane 5b-1 — see
+/// <see cref="ProcessEmailOutboxJob"/> for why the assembly is not load-bearing and why
+/// the file sits under <c>Contracts/</c>.
+/// </remarks>
 [DisableConcurrentExecution(timeoutInSeconds: 300)]
 public class CleanupEmailOutboxJob(
     IEmailOutboxRetention outbox,
