+++ src/Sections/Humans.Email/Docs/Email.md
@@ -15,7 +15,7 @@ Transactional email outbox: queue, render, deliver, retry, pause/resume. Backs c
 - An **Outbox Message** is a single queued email record with recipient, subject, rendered HTML body, status, retry metadata, and optional links to `User` / `CampaignGrant` / `ShiftSignup`.
 - The **Outbox Pause Flag** is a `SystemSetting` row keyed `IsEmailSendingPaused` that, when `"true"`, causes `ProcessEmailOutboxJob` to skip all delivery attempts on its next tick. Resuming flips it back to `"false"`.
 - **Email Body Composition** happens entirely inside the section: `IEmailBodyComposer` / `BrandedEmailBodyComposer` are internal, as are `IEmailRenderer` / `EmailRenderer` and the two transports. Business code outside the section never sees any of them — it builds an `EmailMessage` through `IEmailMessageFactory` and hands it to `IEmailService`, both on `Humans.Email.Contracts`.
-- **Delivery** is performed by `EmailOutboxProcessor` (section) via `IEmailTransport` (`SmtpEmailTransport` in prod, `StubEmailTransport` in dev/test). `ProcessEmailOutboxJob` (Infrastructure) is the Hangfire scheduler shim that calls it through `IEmailOutboxProcessor`. `IImmediateOutboxProcessor` (`HangfireImmediateOutboxProcessor`, Infrastructure) is the trigger for time-sensitive templates that need to fire the next job run immediately rather than wait for the recurring tick.
+- **Delivery** is performed by `EmailOutboxProcessor` (section) via `IEmailTransport` (`SmtpEmailTransport` in prod, `StubEmailTransport` in dev/test). `ProcessEmailOutboxJob` (`Humans.Email/Contracts/`) is the Hangfire scheduler shim that calls it through `IEmailOutboxProcessor`. `IImmediateOutboxProcessor` (`HangfireImmediateOutboxProcessor`, also `Humans.Email/Contracts/`) is the trigger for time-sensitive templates that need to fire the next job run immediately rather than wait for the recurring tick.
 - **One `IEmailService` implementation exists:** `OutboxEmailService` (Application, default — writes to the outbox). DI binds `IEmailService` to `OutboxEmailService`.
 
 ## Data Model
@@ -148,8 +148,8 @@ Per design-rules §8, each `system_settings` key is owned by its consuming secti
   - `IEmailService` + `EmailMessage` — the one transport entry point, called by nine `Humans.Application` services, six `Humans.Infrastructure` jobs and six moved sections.
   - `IEmailMessageFactory` — the typed builders those callers use to construct an `EmailMessage`.
   - `IEmailOutboxServiceRead` + `EmailOutboxMessageDto` — per-human outbox history for Shell's `/Profile/Me/Outbox` and `/Users/Admin/{id}/Outbox`.
-  - `IEmailOutboxProcessor` / `IEmailOutboxRetention` — what `ProcessEmailOutboxJob` / `CleanupEmailOutboxJob` drive.
-  - `IImmediateOutboxProcessor` — on the leaf for the opposite reason: Base *implements* it (`HangfireImmediateOutboxProcessor` enqueues the recurring job) and the section consumes it.
+  - `IEmailOutboxProcessor` / `IEmailOutboxRetention` — what `ProcessEmailOutboxJob` / `CleanupEmailOutboxJob` drive. Both jobs moved into `Humans.Email/Contracts/` at G5 lane 5b-1, so these two have no consumer outside the section any more and could move inward in a later pass.
+  - `IImmediateOutboxProcessor` — was on the leaf for the opposite reason: Base *implemented* it. `HangfireImmediateOutboxProcessor` followed the job into `Humans.Email/Contracts/` at the same lane, so both sides are now section-side.
 - **Section-internal abstractions** (`Humans.Email.Services`): `IEmailOutboxService` (admin surface, consumed only by `EmailController`), `IEmailRenderer` / `EmailRenderer`, `IEmailBodyComposer` / `BrandedEmailBodyComposer`, `IEmailTransport` / `SmtpEmailTransport` / `StubEmailTransport`.
 - **`EmailSettings` stays in `Humans.Infrastructure.Configuration` and is bound in Shell**, not in `Section.Register`: Auth's `MagicLinkUrlBuilder`, Profiles' `UnsubscribeTokenProvider`, `SendReConsentReminderJob` and Shell's `SmtpHealthCheck` all read it. It is Base configuration the section is merely named after.
 - **`EmailOutboxStatus` stays in `Humans.Domain.Enums`.** Campaigns' `CampaignGrant.LatestEmailStatus` and Surveys' `SurveyInvitation` persist it on their own tables, so it is shared Base vocabulary rather than section-internal; its `Enum_EmailOutboxStatus_*` resource keys stay in `SharedResource` with it.
