+++ src/Sections/Humans.Issues/Humans.Issues.csproj
@@ -33,6 +33,8 @@
   </ItemGroup>
 
   <ItemGroup>
+    <!-- Hangfire.Core only: CleanupIssuesJob's [DisableConcurrentExecution]. Hosting stays in Shell. -->
+    <PackageReference Include="Hangfire.Core" />
     <PackageReference Include="Microsoft.EntityFrameworkCore" />
     <PackageReference Include="NodaTime" />
     <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" />
