+++ src/Sections/Humans.Notifications/Contracts/CleanupNotificationsJob.cs
@@ -1,9 +1,8 @@
 using Hangfire;
 using Humans.Application.Interfaces;
-using Humans.Notifications.Contracts;
 using Microsoft.Extensions.Logging;
 
-namespace Humans.Infrastructure.Jobs;
+namespace Humans.Notifications.Contracts;
 
 /// <summary>
 /// Purges old notifications. Runs daily. The retention rule itself — resolved older than
@@ -11,6 +10,14 @@ namespace Humans.Infrastructure.Jobs;
 /// sources — lives inside the Notifications section behind
 /// <see cref="INotificationRetention"/>; this job is the scheduler shim around it.
 /// </summary>
+/// <remarks>
+/// Moved out of <c>Humans.Infrastructure/Jobs</c> at G5 lane 5b-1
+/// (nobodies-collective/Humans#866). The "Hangfire pins a job to its assembly" claim that
+/// kept it in Base was re-measured and is false: <c>RecurringJob.AddOrUpdate&lt;T&gt;(id, …)</c>
+/// rewrites the stored type string on every startup, so the job id is the stable key. It
+/// sits under <c>Contracts/</c> because Shell names the concrete type at registration and
+/// HUM0034 makes every other public type in a section assembly an error.
+/// </remarks>
 [DisableConcurrentExecution(timeoutInSeconds: 300)]
 public class CleanupNotificationsJob(
     INotificationRetention notifications,
