+++ src/Sections/Humans.Notifications/Docs/Notifications.md
@@ -141,7 +141,7 @@ Inbound (other sections → Notifications):
 **Owned tables:** `notifications`, `notification_recipients`
 **Status:** (A) Migrated (peterdrier/Humans PR for issue nobodies-collective/Humans#550, 2026-04-22). Moved into its own project at G5 (nobodies-collective/Humans#866).
 
-- The section is `src/Sections/Humans.Notifications` — everything in it is `internal` except `Section` and `NotificationsResource`. Its cross-section surface is the leaf project `Humans.Notifications.Contracts`: `INotificationEmitter`, `INotificationService`, `INotificationAutoResolve` (the two `ResolveBySource*` calls other sections make), `INotificationRetention` (the nightly purge `CleanupNotificationsJob` drives), and the three enums those signatures name — `NotificationSource`, `NotificationClass`, `NotificationPriority`. It is a project rather than a `Contracts/` folder because most consumers are still in Base. The inbox read models, the meter surface, the recipient resolver and the repository have no consumer outside the section and are internal.
+- The section is `src/Sections/Humans.Notifications` — everything in it is `internal` except `Section`, `NotificationsResource`, and `CleanupNotificationsJob` under `Contracts/` (Shell names the job's concrete type at registration; HUM0034's `Contracts/` carve-out is what lets it be public). Its cross-section surface is the leaf project `Humans.Notifications.Contracts`: `INotificationEmitter`, `INotificationService`, `INotificationAutoResolve` (the two `ResolveBySource*` calls other sections make), `INotificationRetention` (the nightly purge `CleanupNotificationsJob` drives), and the three enums those signatures name — `NotificationSource`, `NotificationClass`, `NotificationPriority`. It is a project rather than a `Contracts/` folder because most consumers are still in Base. The inbox read models, the meter surface, the recipient resolver and the repository have no consumer outside the section and are internal.
 - `INotificationRepository` (impl `src/Sections/Humans.Notifications/Data/NotificationRepository.cs`) is the only non-test file that touches `notifications` / `notification_recipients` via `DbContext`. The repository uses `IDbContextFactory<NotificationsDbContext>` (peeled out of `HumansDbContext` in #858) so it can be registered as **Singleton** while `NotificationsDbContext` remains Scoped.
 - The notification bell is the section's own view component (`ViewComponents/NotificationBellViewComponent`, `internal`). Shell's `_Layout` and `Humans.UI`'s `_AdminLayout` render it with `@await Component.InvokeAsync("NotificationBell")` — by name, across application parts. `Humans.Web/Infrastructure/SectionViewComponentFeatureProvider` is what lets MVC discover a non-public one, the counterpart of `SectionControllerFeatureProvider` for controllers.
 - **DI cycle break — `INotificationEmitter` vs `INotificationService`.** `INotificationEmitter` is the narrow outbound interface (`SendAsync` to a known recipient list). `INotificationService` extends it with `SendToRoleAsync`, which requires `INotificationRecipientResolver` (which itself depends on `IRoleAssignmentService`). Since `IRoleAssignmentService` already depends on the notification surface, it injects the narrower `INotificationEmitter` (implemented by a separate `NotificationEmitter` type, **not** `NotificationService`) to avoid closing the DI cycle.
