+++ tests/Humans.Notifications.Tests/NotificationsArchitectureTests.cs
@@ -125,17 +125,22 @@ public class NotificationsArchitectureTests
     }
 
     [HumansFact]
-    public void NotificationsResourceIsTheOnlyPublicTypeBesidesSection()
+    public void SectionPublicSurfaceIsTheResourceMarkerTheSectionEntryPointAndTheJob()
     {
         // HUM0034 is the build gate; this pins the intent so a Grandfathered escape or a
-        // future carve-out shows up as a test failure too.
+        // future carve-out shows up as a test failure too. It caught CleanupNotificationsJob
+        // arriving at G5 lane 5b-1 (nobodies-collective/Humans#866) — deliberately: the job
+        // sits under Contracts/ because Shell names its concrete type at both its DI
+        // registration and its RecurringJob.AddOrUpdate<T> line, which is HUM0034's
+        // Contracts/ carve-out and not an accident. Anything else new still fails here.
         var publicNames = typeof(Section).Assembly.GetExportedTypes()
             .Select(t => t.Name)
             .Where(n => !n.StartsWith("Baseline", StringComparison.Ordinal))
             .OrderBy(n => n, StringComparer.Ordinal)
             .ToList();
 
-        publicNames.Should().BeEquivalentTo(["NotificationsResource", "Section"]);
+        publicNames.Should().BeEquivalentTo(
+            ["CleanupNotificationsJob", "NotificationsResource", "Section"]);
     }
 
     // ── INotificationRepository ──────────────────────────────────────────────
